Build toolchains in a CentOS6 container
CentOS6/RHEL6 is old, but still used in the EDA industry. Building our
cross-compilation toolchain with the oldest possible distribution allows
us to produce binaries that work on CentOS6/RHEL6 and all more recent
distributions up to today.

Based on earlier work by Sam Elliott, which was based on Holy Build Box.
Unfortunately, HBB version 2.2 (based on CentOS6) does not provide a
version of git with support for HTTPS remotes, which is something we
need to check out further repositories.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>